First working version after zinb loss implementation
