+++ mlruns/0/1dcc9422362042c5b69ab968946c558a/tags/mlflow.source.name
@@ -1,0 +1,1 @@
+/Users/sebastian.birk/opt/anaconda3/envs/autotalkercpu/lib/python3.9/site-packages/ipykernel_launcher.py
